Added the phinx configuration file to gitignore
This file will need to be created and manually edited in order to use
the seeds.

Run the following to generate the file:

``` bash
php vendor/bin/phinx init .
```

Then edit the `paths` section to look like the following:

``` yaml
paths:
    migrations: %%PHINX_CONFIG_DIR%%/config/Migrations
    seeds: %%PHINX_CONFIG_DIR%%/config/Seeds
```

Then change the database configurations to match your environment.